Deploy 24/5

+++ app/Http/Controllers/ProfesorEscribeForo.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\usuarios;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Storage;
 use App\Models\datosForo;
 use App\Models\Foro;
@@ -15,48 +17,75 @@ use App\Models\ProfesorForoGrupo;
 
 class ProfesorEscribeForo extends Controller
 {
-    public function index(Request $request)
+    public function getForoId($idGrupo, $idMateria)
     {
-        return response()->json(ProfesorForoGrupo::where('idMateria', $request->idMateria)
-            ->where('idGrupo', $request->idGrupo)->first());
+        return response()->json(ProfesorForoGrupo::where('idMateria', $idMateria)
+            ->where('idGrupo', $idGrupo)->first());
     }
 
-    public function traerGrupos(Request $request){
-        if($request->ou == 'Profesor'){
-            $request["idProfesor"] = $request->idUsuario;
+    public function traerGrupos(Request $request, $id)
+    {
+        $usuario = usuarios::findOrFail($id);
+        if ($usuario->ou == 'Profesor') {
+            $request["idProfesor"] = $usuario->id;
             return ProfesorGrupo::listarProfesorGrupo($request);
-        }else if ($request->ou == 'Alumno'){
-            return self::traerGruposAlumnos($request);
+        } else if ($usuario->ou == 'Alumno') {
+            return self::traerGruposAlumnos($usuario->id);
         }
     }
 
- 
-    public function traerGruposAlumnos($request){
-        $gruposAlumno=alumnoGrupo::select('idGrupo')->where('idAlumnos', $request->idUsuario)->get();
+
+    public function traerGruposAlumnos($id)
+    {
+        $gruposAlumno = alumnoGrupo::select('idGrupo')->where('idAlumnos', $id)->get();
         return response()->json($gruposAlumno);
     }
 
 
-    public function traerArchivo(Request $request)
+    public function traerArchivo($archivo)
     {
-        return Storage::disk('ftp')->get($request->archivo);
+        if (empty($archivo)) {
+            return response()->json(['error' => 'Archivo no encontrado'], 404);
+        }
+        try {
+            return Storage::disk('ftp')->get($archivo);
+        } catch (\Exception $e) {
+            return response()->json(['error' => 'Archivo no encontrado'], 404);
+        }
     }
 
-    public function show(Request $request)
+    public function getAllPublicaciones($idGrupo, $idUsuario, $limit)
     {
-        if ($request->idMateria) {
-            if ($request->ou == 'Profesor') {
-                return  self::traerPublicacionesProfesorMateria($request);
-            } else if ($request->ou == 'Alumno') {
-                return self::traerPublicacionesAlumnoMateria($request);
-            }
-        } else {
-            if ($request->ou == 'Profesor') {
-                return  self::traerPublicacionesProfesor($request);
-            } else if ($request->ou == 'Alumno') {
-                return self::traerPublicacionesAlumno($request);
-            }
+        $usuario = usuarios::findOrFail($idUsuario);
+        $data = [
+            'idUsuario' => $idUsuario,
+            'idGrupo' => $idGrupo,
+            'limit' => $limit
+        ];
+
+        if ($usuario->ou == 'Profesor') {
+            return self::traerPublicacionesProfesor($data);
+        } else if ($usuario->ou == 'Alumno') {
+            return self::traerPublicacionesAlumno($data);
         }
+    }
+
+    public function getAllPublicacionesMateria($idGrupo, $idUsuario, $idMateria, $limit)
+    {
+        $usuario = usuarios::findOrFail($idUsuario);
+        $data = [
+            'idUsuario' => $idUsuario,
+            'idGrupo' => $idGrupo,
+            'idMateria' => $idMateria,
+            'limit' => $limit,
+        ];
+        if ($usuario->ou == 'Profesor') {
+            return self::traerPublicacionesProfesorMateria($data);
+        } else if ($usuario->ou == 'Alumno') {
+            return self::traerPublicacionesAlumnoMateria($data);
+        }
+
+        return response()->json([]);
     }
 
 
@@ -77,8 +106,12 @@ class ProfesorEscribeForo extends Controller
             $arrayImagenes = array();
             $postAuthor = $p->postAuthor;
             $imgPerfil = $this->getImagenPefil($postAuthor);
+            if (!App::environment(['testing'])) {
+                $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
+            } else {
+                $img = $imgPerfil[0]->imagen_perfil;
+            }
 
-            $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
 
 
             foreach ($peticionSQLFiltrada as $p2) {
@@ -87,7 +120,12 @@ class ProfesorEscribeForo extends Controller
                 if ($resultado) {
                     array_push($arrayArchivos, $p2->archivo);
                 } else {
-                    array_push($arrayImagenes, base64_encode(Storage::disk('ftp')->get($p2->archivo)));
+                    if (!App::environment(['testing'])) {
+                        array_push($arrayImagenes, base64_encode(Storage::disk('ftp')->get($p2->archivo)));
+                    } else {
+                        array_push($arrayImagenes, $p2->archivo);
+                    }
+
                 }
             }
 
@@ -183,8 +221,12 @@ class ProfesorEscribeForo extends Controller
             $postAuthor = $p->postAuthor;
 
             $imgPerfil = $this->getImagenPefil($postAuthor);
+            if (!App::environment(['testing'])) {
+                $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
+            } else {
+                $img = $imgPerfil[0]->imagen_perfil;
+            }
 
-            $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
 
             foreach ($peticionSQLFiltrada as $p2) {
 
@@ -235,8 +277,12 @@ class ProfesorEscribeForo extends Controller
             $postAuthor = $p->postAuthor;
 
             $imgPerfil = $this->getImagenPefil($postAuthor);
-
-            $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
+            if (!App::environment(['testing'])) {
+                $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
+            }else{
+                $img = $imgPerfil[0]->imagen_perfil;
+            }
+          
 
             foreach ($peticionSQLFiltrada as $p2) {
 
@@ -271,47 +317,67 @@ class ProfesorEscribeForo extends Controller
         return response()->json($dataResponse);
     }
 
-    public function store (Request $request){
+    public function store(Request $request)
+    {
+        $request->validate(
+            [
+                'idGrupo' => 'required',
+                'idMateria' => 'required',
+                'idUsuario' => 'required',
+                'mensaje' => 'string',
+                'archivos' => 'array | nullable',
+                'nombresArchivo' => 'array | nullable'
+            ]
+        );
 
-        $this->agregarDatosForo($request);
 
-        $idDatos = DB::table('datosForo')->orderBy('created_at', 'desc')->limit(1)->get('id');
-
-        if ($request->archivos){
-            for ($i=0; $i < count($request->nombresArchivo); $i++){
-                $this->subirArchivoForo($request, $i, $idDatos[0]);
+        $datoForo = $this->agregarDatosForo($request);
+        if ($request->archivos) {
+            for ($i = 0; $i < count($request->nombresArchivo); $i++) {
+                $this->subirArchivoForo($request, $i, $datoForo->id);
             }
-    }
+        }
 
-        RegistrosController::store("PUBLICACION FORO",$request->header('token'),"CREATE","");
+        RegistrosController::store("PUBLICACION FORO", $request->header('token'), "CREATE", "");
         return response()->json(['status' => 'Success'], 200);
     }
 
-
-    public function destroy(Request $request)
+    public function agregarDatosForo(Request $request)
     {
+        $foro = ProfesorForoGrupo::where('idGrupo', $request->idGrupo)->where('idMateria', $request->idMateria)->first();
+        $datosForo = new datosForo;
+        $datosForo->idForo = $foro->idForo;
+        $datosForo->idUsuario = $request->idUsuario;
+        $datosForo->mensaje = $request->mensaje;
+        $datosForo->save();
 
-        $postForo = datosForo::where('id', $request->id)->first();
-        $arhivosForo = archivosForo::where('idDato', $request->id)->get();
+        return $datosForo;
+    }
+
+
+    public function destroy(Request $request, $id)
+    {
+        $postForo = datosForo::findOrFail($id);
+        $arhivosForo = archivosForo::where('idDato', $id)->get();
         foreach ($arhivosForo as $p) {
-            Storage::disk('ftp')->delete($p->nombreArchivo);
+            if (!App::environment(['testing'])) {
+                Storage::disk('ftp')->delete($p->nombreArchivo);
+            }
+          
             $arhivosId = archivosForo::where('id', $p->id)->first();
             $arhivosId->delete();
         }
         try {
             $postForo->delete();
-            RegistrosController::store("PUBLICACION FORO",$request->header('token'),"DELETE","");
+            RegistrosController::store("PUBLICACION FORO", $request->header('token'), "DELETE", "");
             return response()->json(['status' => 'Success'], 200);
         } catch (\Throwable $th) {
             return response()->json(['status' => 'Bad Request'], 400);
         }
     }
 
-    /**
-     * @param $request
-     * @return \Illuminate\Support\Collection
-     */
-    public function getPublicacionesForoForProfesor($request): \Illuminate\Support\Collection
+
+    public function getPublicacionesForoForProfesor($request)
     {
         $peticionSQL = DB::table('profesor_estan_grupo_foro')
             ->select('datosForo.id AS id', 'datosForo.idForo AS idForo', 'profesor_estan_grupo_foro.idGrupo', 'materias.nombre AS materia', 'datosForo.idUsuario AS idUsuario', 'usuarios.nombre AS nombreAutor', 'datosForo.mensaje AS mensaje', 'datosForo.created_at AS fecha', 'datosForo.idUsuario as postAuthor')
@@ -320,21 +386,18 @@ class ProfesorEscribeForo extends Controller
             ->join('grupos_tienen_profesor', 'grupos_tienen_profesor.idGrupo', '=', 'profesor_estan_grupo_foro.idGrupo')
             ->join('usuarios', 'usuarios.id', '=', 'datosForo.idUsuario')
             ->join('materias', 'materias.id', '=', 'profesor_estan_grupo_foro.idMateria')
-            ->where('profesor_estan_grupo_foro.idProfesor', $request->idUsuario)
-            ->where('profesor_estan_grupo_foro.idGrupo', $request->idGrupo)
-            ->where('grupos_tienen_profesor.deleted_at', NULL)
+            ->where('profesor_estan_grupo_foro.idProfesor', $request['idUsuario'])
+            ->where('profesor_estan_grupo_foro.idGrupo', $request['idGrupo'])
+
             ->orderBy('id', 'desc')
-            ->take($request->limit)
+            ->take($request['limit'])
             ->distinct()
             ->get();
         return $peticionSQL;
     }
 
-    /**
-     * @param $p
-     * @return \Illuminate\Support\Collection
-     */
-    public function getArchivosForo($p): \Illuminate\Support\Collection
+
+    public function getArchivosForo($p)
     {
         $peticionSQLFiltrada = DB::table('archivos_foro')
             ->select('nombreArchivo AS archivo')
@@ -344,11 +407,8 @@ class ProfesorEscribeForo extends Controller
         return $peticionSQLFiltrada;
     }
 
-    /**
-     * @param $postAuthor
-     * @return \Illuminate\Support\Collection
-     */
-    public function getImagenPefil($postAuthor): \Illuminate\Support\Collection
+
+    public function getImagenPefil($postAuthor)
     {
         $imgPerfil = DB::table('usuarios')
             ->select('imagen_perfil')
@@ -357,107 +417,80 @@ class ProfesorEscribeForo extends Controller
         return $imgPerfil;
     }
 
-    /**
-     * @param $request
-     * @return \Illuminate\Support\Collection
-     */
-    public function getPublicacionesForoMateriaForProfesor($request): \Illuminate\Support\Collection
+
+    public function getPublicacionesForoMateriaForProfesor($request)
     {
         $peticionSQL = DB::table('profesor_estan_grupo_foro')
             ->select('datosForo.id AS id', 'datosForo.idForo AS idForo', 'profesor_estan_grupo_foro.idGrupo', 'materias.nombre as materia', 'datosForo.idUsuario AS idUsuario', 'usuarios.nombre AS nombreAutor', 'datosForo.mensaje AS mensaje', 'datosForo.created_at AS fecha', 'datosForo.idUsuario as postAuthor')
             ->join('datosForo', 'datosForo.idForo', '=', 'profesor_estan_grupo_foro.idForo')
             ->join('usuarios', 'usuarios.id', '=', 'datosForo.idUsuario')
             ->join('materias', 'materias.id', '=', 'profesor_estan_grupo_foro.idMateria')
-            ->where('profesor_estan_grupo_foro.idProfesor', $request->idUsuario)
-            ->where('profesor_estan_grupo_foro.idMateria', $request->idMateria)
-            ->where('profesor_estan_grupo_foro.idGrupo', $request->idGrupo)
+            ->where('profesor_estan_grupo_foro.idProfesor', $request['idUsuario'])
+            ->where('profesor_estan_grupo_foro.idMateria', $request['idMateria'])
+            ->where('profesor_estan_grupo_foro.idGrupo', $request['idGrupo'])
             ->orderBy('id', 'desc')
-            ->take($request->limit)
+            ->take($request['limit'])
             ->get();
         return $peticionSQL;
     }
 
-    /**
-     * @param $request
-     * @return \Illuminate\Support\Collection
-     */
-    public function getIdGrupoAlumno($request): \Illuminate\Support\Collection
+
+    public function getIdGrupoAlumno($request)
     {
         $idGrupo = DB::table('alumnos_pertenecen_grupos')
             ->select('alumnos_pertenecen_grupos.idGrupo AS idGrupo')
             ->where('alumnos_pertenecen_grupos.idAlumnos', $request->idUsuario)
-            ->where('alumnos_pertenecen_grupos.deleted_at', NULL)
             ->get();
         return $idGrupo;
     }
 
-    /**
-     * @param $idGrupo
-     * @param $request
-     * @return \Illuminate\Support\Collection
-     */
-    public function getPublicacionesForoForGrupo($request): \Illuminate\Support\Collection
+
+    public function getPublicacionesForoForGrupo($request)
     {
         $peticionSQL = DB::table('profesor_estan_grupo_foro')
             ->select('datosForo.id AS id', 'datosForo.idForo AS idForo', 'profesor_estan_grupo_foro.idGrupo', 'materias.nombre as materia', 'datosForo.idUsuario AS idUsuario', 'usuarios.nombre AS nombreAutor', 'datosForo.mensaje AS mensaje', 'datosForo.created_at AS fecha', 'datosForo.idUsuario as postAuthor')
             ->join('materias', 'materias.id', '=', 'profesor_estan_grupo_foro.idMateria')
             ->join('datosForo', 'datosForo.idForo', '=', 'profesor_estan_grupo_foro.idForo')
             ->join('usuarios', 'usuarios.id', '=', 'datosForo.idUsuario')
-            ->where('profesor_estan_grupo_foro.idGrupo', $request->idGrupo)
+            ->where('profesor_estan_grupo_foro.idGrupo', $request['idGrupo'])
             ->orderBy('id', 'desc')
-            ->take($request->limit)
+            ->take($request['limit'])
             ->distinct()
             ->get();
         return $peticionSQL;
     }
 
-    /**
-     * @param $idGrupo
-     * @param $request
-     * @return \Illuminate\Support\Collection
-     */
-    public function getPublicacionesForoMateriaForGrupo($request): \Illuminate\Support\Collection
+
+    public function getPublicacionesForoMateriaForGrupo($request)
     {
         $peticionSQL = DB::table('profesor_estan_grupo_foro')
             ->select('datosForo.id AS id', 'datosForo.idForo AS idForo', 'profesor_estan_grupo_foro.idGrupo', 'materias.nombre as materia', 'datosForo.idUsuario AS idUsuario', 'usuarios.nombre AS nombreAutor', 'datosForo.mensaje AS mensaje', 'datosForo.created_at AS fecha', 'datosForo.idUsuario as postAuthor')
             ->join('materias', 'materias.id', '=', 'profesor_estan_grupo_foro.idMateria')
             ->join('datosForo', 'datosForo.idForo', '=', 'profesor_estan_grupo_foro.idForo')
             ->join('usuarios', 'usuarios.id', '=', 'datosForo.idUsuario')
-            ->where('profesor_estan_grupo_foro.idGrupo', $request->idGrupo)
-            ->where('profesor_estan_grupo_foro.idMateria', $request->idMateria)
+            ->where('profesor_estan_grupo_foro.idGrupo', $request['idGrupo'])
+            ->where('profesor_estan_grupo_foro.idMateria', $request['idMateria'])
             ->orderBy('id', 'desc')
-            ->take($request->limit)
+            ->take($request['limit'])
             ->get();
         return $peticionSQL;
     }
 
-    /**
-     * @param Request $request
-     * @return void
-     */
-    public function agregarDatosForo(Request $request): void
-    {
-        $datosForo = new datosForo;
-        $datosForo->idForo = $request->idForo;
-        $datosForo->idUsuario = $request->idUsuario;
-        $datosForo->mensaje = $request->mensaje;
-        $datosForo->save();
-    }
 
-    /**
-     * @param Request $request
-     * @param int $i
-     * @param $idDatos
-     * @return void
-     * @throws \Exception
-     */
-    public function subirArchivoForo(Request $request, int $i, $idDatos): void
+
+
+
+    public function subirArchivoForo(Request $request, int $i, $idDatos)
     {
+        $foro = ProfesorForoGrupo::where('idGrupo', $request->idGrupo)->where('idMateria', $request->idMateria)->first();
         $nombreArchivo = random_int(0, 1000000) . "_" . $request->nombresArchivo[$i];
-        Storage::disk('ftp')->put($nombreArchivo, fopen($request->archivos[$i], 'r+'));
+        if (!App::environment(['testing'])) {
+            Storage::disk('ftp')->put($nombreArchivo, fopen($request->archivos[$i], 'r+'));
+        }
+       
         $archivosForo = new archivosForo;
-        $archivosForo->idDato = $idDatos->id;
-        $archivosForo->idForo = $request->idForo;
+        $archivosForo->idDato = $idDatos;
+        $archivosForo->idForo = $foro->idForo;
         $archivosForo->nombreArchivo = $nombreArchivo;
         $archivosForo->save();
     }
